(nvim) Add workspace library config
Add workspace.library config so that lua-language-server knows about
different definitions and completions that aren't in the workspace
scope. There is likely a better way to do this that I don't have time to
look into right now. Reference the [configuration guide] when attempting
to make this list more dynamic and flexible to change later.

[configuration guide]: https://github.com/LuaLS/lua-language-server/wiki/Configuration-File