Let gitmessage use `#` md header syntax
The default line start comment character (the character that tells `git`
to discard the line from the commit message) is `#` which conflicts with
using this character to specify header levels in markdown syntax.

This can be avoided by using a sequence of `----` or `======` under the
heading, but this only provides two heading levels.

This change modifies the commentChar to `[` and updates the empty git
commit message accordingly.