Setup .localrc concept
Sometimes there are local workstation specific configurations that I
don't want to persist to my dotfiles but I do want to keep a reference
of this concept and integrate it into my main rc config.

This makes sure there is a ~/.localrc file available and references it
from profile and zshrc configs.