Test out using oil instead of vim-vinegar
I think vim-vinegar is great but there are a few features that I feel
will be helpful by adding the "more" featured oil.nvim. The ability to
use vim commands to mass rename and move feels good as opposed to
"mark-and-target-and-command" flow from netrw.

Will try for a while and then re-configure