Use `shortlog` for finding coauthors
We can get a succinct listing of authors of git commits with `shortlog`
instead of searching back a certain number of commits. Don't know how
this performs on large commit trees but :shrug:

Ran into an interesting issue where `shortlog` will look to analyze
STDIN when not in a terminal so had to execute a `sh -c ""` for piping
and `cut`. Obviously not the most portable but works for me.

Finally, can't call `vim.notify` in a system callback so have to use
`vim.schedule` to add the UI update.